refactor(userspace/libsinsp): final renaming pass to filtercheck sources
