+++ userspace/libsinsp/CMakeLists.txt
@@ -87,23 +87,23 @@ set(SINSP_SOURCES
 	fdinfo.cpp
 	filter.cpp
 	container.cpp
-	filtercheck_container.cpp
-	filtercheck_event.cpp
-	filtercheck_evtin.cpp
-	filtercheck_fd.cpp
-	filtercheck_fdlist.cpp
-	filtercheck_fspath.cpp
-	filtercheck_gen_event.cpp
-	filtercheck_group.cpp
-	filtercheck_k8s.cpp
-	filtercheck_mesos.cpp
-	filtercheck_rawstring.cpp
-	filtercheck_reference.cpp
-	filtercheck_syslog.cpp
-	filtercheck_thread.cpp
-	filtercheck_tracer.cpp
-	filtercheck_user.cpp
-	filtercheck_utils.cpp
+	sinsp_filtercheck_container.cpp
+	sinsp_filtercheck_event.cpp
+	sinsp_filtercheck_evtin.cpp
+	sinsp_filtercheck_fd.cpp
+	sinsp_filtercheck_fdlist.cpp
+	sinsp_filtercheck_fspath.cpp
+	sinsp_filtercheck_gen_event.cpp
+	sinsp_filtercheck_group.cpp
+	sinsp_filtercheck_k8s.cpp
+	sinsp_filtercheck_mesos.cpp
+	sinsp_filtercheck_rawstring.cpp
+	sinsp_filtercheck_reference.cpp
+	sinsp_filtercheck_syslog.cpp
+	sinsp_filtercheck_thread.cpp
+	sinsp_filtercheck_tracer.cpp
+	sinsp_filtercheck_user.cpp
+	sinsp_filtercheck_utils.cpp
 	filter_check_list.cpp
 	gen_filter.cpp
 	http_parser.c
