+++ userspace/libsinsp/filterchecks.h
@@ -22,20 +22,20 @@ limitations under the License.
 // reasons, as all filterchecks used to be defined in this file
 
 #include "sinsp_filtercheck.h"
-#include "filtercheck_container.h"
-#include "filtercheck_event.h"
-#include "filtercheck_evtin.h"
-#include "filtercheck_fd.h"
-#include "filtercheck_fdlist.h"
-#include "filtercheck_fspath.h"
-#include "filtercheck_gen_event.h"
-#include "filtercheck_group.h"
-#include "filtercheck_k8s.h"
-#include "filtercheck_mesos.h"
-#include "filtercheck_rawstring.h"
-#include "filtercheck_reference.h"
-#include "filtercheck_syslog.h"
-#include "filtercheck_thread.h"
-#include "filtercheck_tracer.h"
-#include "filtercheck_user.h"
-#include "filtercheck_utils.h"
+#include "sinsp_filtercheck_container.h"
+#include "sinsp_filtercheck_event.h"
+#include "sinsp_filtercheck_evtin.h"
+#include "sinsp_filtercheck_fd.h"
+#include "sinsp_filtercheck_fdlist.h"
+#include "sinsp_filtercheck_fspath.h"
+#include "sinsp_filtercheck_gen_event.h"
+#include "sinsp_filtercheck_group.h"
+#include "sinsp_filtercheck_k8s.h"
+#include "sinsp_filtercheck_mesos.h"
+#include "sinsp_filtercheck_rawstring.h"
+#include "sinsp_filtercheck_reference.h"
+#include "sinsp_filtercheck_syslog.h"
+#include "sinsp_filtercheck_thread.h"
+#include "sinsp_filtercheck_tracer.h"
+#include "sinsp_filtercheck_user.h"
+#include "sinsp_filtercheck_utils.h"
