+++ userspace/libsinsp/sinsp_filtercheck_event.cpp
@@ -18,7 +18,7 @@ limitations under the License.
 
 #include <math.h>
 
-#include "filtercheck_event.h"
+#include "sinsp_filtercheck_event.h"
 #include "sinsp.h"
 #include "sinsp_int.h"
 #include "filter.h"
