+++ userspace/libsinsp/sinsp_filtercheck_fdlist.cpp
@@ -16,7 +16,7 @@ limitations under the License.
 
 */
 
-#include "filtercheck_fdlist.h"
+#include "sinsp_filtercheck_fdlist.h"
 #include "sinsp.h"
 #include "sinsp_int.h"
 #include "filter.h"
