+++ userspace/libsinsp/sinsp_filtercheck_fspath.h
