+++ userspace/libsinsp/sinsp_filtercheck_group.cpp
@@ -16,7 +16,7 @@ limitations under the License.
 
 */
 
-#include "filtercheck_group.h"
+#include "sinsp_filtercheck_group.h"
 #include "sinsp.h"
 #include "sinsp_int.h"
 #include "filter.h"
