+++ userspace/libsinsp/sinsp_filtercheck_k8s.h
