+++ userspace/libsinsp/sinsp_filtercheck_mesos.cpp
@@ -16,7 +16,7 @@ limitations under the License.
 
 */
 
-#include "filtercheck_mesos.h"
+#include "sinsp_filtercheck_mesos.h"
 #include "sinsp.h"
 #include "sinsp_int.h"
 #include "filter.h"
