+++ userspace/libsinsp/sinsp_filtercheck_rawstring.cpp
@@ -16,7 +16,7 @@ limitations under the License.
 
 */
 
-#include "filtercheck_rawstring.h"
+#include "sinsp_filtercheck_rawstring.h"
 #include "sinsp.h"
 #include "sinsp_int.h"
 #include "filter.h"
