+++ userspace/libsinsp/sinsp_filtercheck_rawstring.h
