+++ userspace/libsinsp/sinsp_filtercheck_syslog.cpp
@@ -16,7 +16,7 @@ limitations under the License.
 
 */
 
-#include "filtercheck_syslog.h"
+#include "sinsp_filtercheck_syslog.h"
 #include "sinsp.h"
 #include "sinsp_int.h"
 #include "filter.h"
