+++ userspace/libsinsp/sinsp_filtercheck_syslog.h
