+++ userspace/libsinsp/sinsp_filtercheck_tracer.cpp
@@ -18,8 +18,8 @@ limitations under the License.
 
 #include <math.h>
 
-#include "filtercheck_tracer.h"
-#include "filtercheck_reference.h"
+#include "sinsp_filtercheck_tracer.h"
+#include "sinsp_filtercheck_reference.h"
 #include "sinsp.h"
 #include "sinsp_int.h"
 #include "filter.h"
