+++ userspace/libsinsp/sinsp_filtercheck_utils.cpp
@@ -16,7 +16,7 @@ limitations under the License.
 
 */
 
-#include "filtercheck_utils.h"
+#include "sinsp_filtercheck_utils.h"
 #include "sinsp.h"
 #include "sinsp_int.h"
 #include "filter.h"
