+++ userspace/libsinsp/sinsp_filtercheck_utils.h
